Update PimaIndian.R
Load the Pima Indians Diabetes data from the mlbench package.

i) Apply kNN and LDA to the Pima Indians Diabetes data: randomly split the data to a training set (70%) and a test set (30%) and repeat the random split 20 times.
For kNN, repeat 5-fold cross-validation five times to choose k from (3, 5, 7, 9, 11, 13, 15, 17, 19, 21). Use AUC as the metric to choose k, i.e. choose k with the largest AUC.
Record the 20 AUC values of kNN and LDA in two vectors. 

ii) For the first random split, draw the ROC curves of kNN and LDA for this split on
one plot. 

iii) Draw two boxplots based on the 50 AUC values of kNN and LDA obtained from the 20 random splits.