OrangeJuiceData(SVM)
Use the OJ data from the ISLR package. This data contain 1070 purchases information to study which orange juice a customer would buy. 

The Purchase variable is a factor with levels CH and MM indicating whether the customer purchased Citrus Hill or Minute Maid Orange Juice. 
17 features of the customers and products are recorded. 
The details of this dataset can be found at: https://rdrr.io/cran/ISLR/man/OJ.html.

Split the data into a training set (70%) and a test set (30%).
(1) Fitasupportvectorclassifiertothetrainingdatabytuningthecostfrom(0.01,0.1,1,10). Compute the test error rate using the tuned value for cost. 
(2) Fit a support vector machine with a radial kernel to the training data by tuning the cost from (0.01,0.1,1,10). 
      Use the default value of gamma. 
      Compute the test error rate using the tuned value for the cost.
(3) Fit a support vector machine with a polynomial kernel to the training data by tuning the cost from (0.01, 0.1, 1, 10). Use degree=2. 
      Compute the test error rate using the tuned value for the cost.
(4) Draw one plot with three ROC curves for the test predictions in (1), (2), and (3). Comment on the plot. 
(5) Fit a decision tree to the training data with an optimal tree size determined by cross-validation. 
      Create a plot of the pruned tree and interpret the results. 
      If cross- validation does not lead to selection of a pruned tree, then create a pruned tree with five terminal nodes. 
      Compute the test error rate of the pruned tree. 
(6) Fit a random forest to the training data with the following mtry values: 1, 2, 3, 4, 5 and 6. 
     Compute the test error rates and comment on the results. 
     Create a plot showing variable importance for the model with the best test error.
(7) Draw one plot with two ROC curves for the test predictions in (5) and (6).